refactor & update tests
