+++ package.json
@@ -30,60 +30,61 @@
   "author": "Defra DDTS",
   "contributors": [
     "Amirs Ishmuhametovs <amirs.ishmuhametovs@defra.gov.uk>",
-    "Fay Toward <fay.toward@defra.gov.uk>"
+    "Fay Toward <fay.toward@defra.gov.uk>",
+    "Rana Salem <rana.salem@defra.gov.uk>"
   ],
   "license": "OGL-UK-3.0",
   "dependencies": {
-    "@defra/hapi-tracing": "^1.0.0",
-    "@elastic/ecs-pino-format": "^1.5.0",
-    "@hapi/catbox-memory": "^6.0.2",
-    "@hapi/catbox-redis": "^7.0.2",
-    "@hapi/hapi": "^21.3.12",
-    "@hapi/inert": "^7.1.0",
-    "@hapi/vision": "^7.0.3",
-    "@hapi/yar": "^11.0.2",
-    "aws-embedded-metrics": "^4.2.0",
-    "convict": "^6.2.4",
-    "cssnano": "^7.0.6",
-    "cssnano-preset-default": "^7.0.6",
-    "date-fns": "^4.1.0",
-    "govuk-frontend": "^5.7.1",
-    "hapi-pino": "^12.1.0",
-    "hapi-pulse": "^3.0.1",
-    "http-status-codes": "^2.3.0",
-    "https-proxy-agent": "^7.0.6",
-    "ioredis": "^5.4.1",
-    "nunjucks": "^3.2.4",
-    "pino": "^9.5.0",
-    "pino-pretty": "^13.0.0",
-    "standard": "^17.1.2",
-    "undici": "^6.21.1"
+    "@defra/hapi-tracing": "1.0.0",
+    "@elastic/ecs-pino-format": "1.5.0",
+    "@hapi/catbox-memory": "6.0.2",
+    "@hapi/catbox-redis": "7.0.2",
+    "@hapi/hapi": "21.3.12",
+    "@hapi/inert": "7.1.0",
+    "@hapi/vision": "7.0.3",
+    "@hapi/yar": "11.0.2",
+    "aws-embedded-metrics": "4.2.0",
+    "convict": "6.2.4",
+    "cssnano": "7.0.6",
+    "cssnano-preset-default": "7.0.6",
+    "date-fns": "4.1.0",
+    "govuk-frontend": "5.7.1",
+    "hapi-pino": "12.1.0",
+    "hapi-pulse": "3.0.1",
+    "http-status-codes": "2.3.0",
+    "https-proxy-agent": "7.0.6",
+    "ioredis": "5.4.1",
+    "nunjucks": "3.2.4",
+    "pino": "9.5.0",
+    "pino-pretty": "13.0.0",
+    "standard": "17.1.2",
+    "undici": "6.21.1"
   },
   "devDependencies": {
-    "@types/convict": "^6.1.6",
-    "@types/hapi__catbox-memory": "^4.1.8",
-    "@types/jest": "^29.5.14",
+    "@types/convict": "6.1.6",
+    "@types/hapi__catbox-memory": "4.1.8",
+    "@types/jest": "29.5.14",
     "@types/node": "22.9.0",
-    "@types/nunjucks": "^3.2.6",
-    "@types/webpack-assets-manifest": "^5.1.4",
-    "autoprefixer": "^10.4.20",
-    "clean-webpack-plugin": "^4.0.0",
-    "copy-webpack-plugin": "^12.0.2",
-    "jest": "^29.7.0",
-    "jest-environment-jsdom": "^29.7.0",
+    "@types/nunjucks": "3.2.6",
+    "@types/webpack-assets-manifest": "5.1.4",
+    "autoprefixer": "10.4.20",
+    "clean-webpack-plugin": "4.0.0",
+    "copy-webpack-plugin": "12.0.2",
+    "jest": "29.7.0",
+    "jest-environment-jsdom": "29.7.0",
     "jest-fetch-mock": "3.0.3",
-    "nodemon": "^3.1.7",
-    "npm-run-all": "^4.1.5",
-    "postcss-load-config": "^6.0.1",
-    "postcss-loader": "^8.1.1",
-    "sass-embedded": "^1.80.6",
-    "sass-loader": "^16.0.3",
-    "source-map-loader": "^5.0.0",
-    "stylelint": "^16.10.0",
-    "stylelint-config-gds": "^2.0.0",
-    "terser-webpack-plugin": "^5.3.10",
-    "webpack": "^5.96.1",
-    "webpack-assets-manifest": "^5.2.1",
-    "webpack-cli": "^5.1.4"
+    "nodemon": "3.1.7",
+    "npm-run-all": "4.1.5",
+    "postcss-load-config": "6.0.1",
+    "postcss-loader": "8.1.1",
+    "sass-embedded": "1.80.6",
+    "sass-loader": "16.0.3",
+    "source-map-loader": "5.0.0",
+    "stylelint": "16.10.0",
+    "stylelint-config-gds": "2.0.0",
+    "terser-webpack-plugin": "5.3.10",
+    "webpack": "5.96.1",
+    "webpack-assets-manifest": "5.2.1",
+    "webpack-cli": "5.1.4"
   }
 }
