+++ src/config/nunjucks.js
@@ -1,0 +1,16 @@
+import { isDevelopment } from '../constants/environments.js'
+
+export const nunjucksConfig = {
+  nunjucks: {
+    watch: {
+      doc: 'Reload templates when they are changed.',
+      format: Boolean,
+      default: isDevelopment
+    },
+    noCache: {
+      doc: 'Use a cache and recompile templates each time',
+      format: Boolean,
+      default: isDevelopment
+    }
+  }
+}
