+++ src/constants/environments.js
@@ -1,3 +1,3 @@
-export const isProduction = process.env.NODE_ENV === 'production'
-export const isTest = process.env.NODE_ENV === 'test'
-export const isDevelopment = process.env.NODE_ENV === 'development'
+export const isProduction = process.env.NODE_ENV = 'production'
+export const isTest = process.env.NODE_ENV = 'test'
+export const isDevelopment = process.env.NODE_ENV = 'development'
