+++ src/plugins/request-tracing.js
@@ -1,7 +1,7 @@
-import { config } from '../config/config.js'
+import { config } from '../config/index.js'
 import { tracing } from '@defra/hapi-tracing'
 
 export const requestTracing = {
   plugin: tracing.plugin,
-  options: { tracingHeader: config.get('tracing.header') }
+  options: { tracingHeader: config.get('server.tracing.header') }
 }
