+++ src/plugins/template-renderer/context.js
@@ -1,20 +1,20 @@
 import path from 'node:path'
 import { readFileSync } from 'node:fs'
 
-import { config } from '../../config/config.js'
+import { config } from '../../config/index.js'
 import { createLogger } from '../../utils/logger.js'
 import { getNavigationItems } from '../../config/navigation-items.js'
 
 const logger = createLogger()
-const assetPath = config.get('assetPath')
+const assetPath = config.get('server.assetPath')
 const manifestPath = path.join(
-  config.get('root'),
+  config.get('server.root'),
   '.public/assets-manifest.json'
 )
 
 let webpackManifest
 
-export function context (request) {
+export const context = (request) => {
   if (!webpackManifest) {
     try {
       webpackManifest = JSON.parse(readFileSync(manifestPath, 'utf-8'))
@@ -25,7 +25,7 @@ export function context (request) {
 
   return {
     assetPath: `${assetPath}/assets`,
-    serviceName: config.get('serviceName'),
+    serviceName: config.get('server.serviceName'),
     serviceUrl: '/',
     breadcrumbs: [],
     navigation: getNavigationItems(request),
