+++ src/plugins/template-renderer/filters/format-currency.js
@@ -1,4 +1,4 @@
-export function formatCurrency (value, locale = 'en-GB', currency = 'GBP') {
+export const formatCurrency = (value, locale = 'en-GB', currency = 'GBP') => {
   const formatter = new Intl.NumberFormat(locale, {
     style: 'currency',
     currency
