+++ src/plugins/template-renderer/filters/format-date.js
@@ -1,6 +1,6 @@
 import { format, isDate, parseISO } from 'date-fns'
 
-export function formatDate (value, formattedDateStr = 'EEE do MMMM yyyy') {
+export const formatDate = (value, formattedDateStr = 'EEE do MMMM yyyy') => {
   const date = isDate(value) ? value : parseISO(value)
 
   return format(date, formattedDateStr)
