+++ src/plugins/template-renderer/filters/to-error-list.js
@@ -1,4 +1,4 @@
-export function toErrorList (errors) {
+export const toErrorList = (errors) => {
   if (!errors) {
     return []
   }
