+++ src/routes/errors/index.js
@@ -1,8 +1,8 @@
 import { serviceUnavailable } from './service-unavailable.js'
 import { pageNotFound } from './page-not-found.js'
 import { serviceProblem } from './service-problem.js'
-import { config } from '../../config/config.js'
+import { config } from '../../config/index.js'
 
-export const errors = config.get('allowErrorViews')
+export const errors = config.get('server.allowErrorViews')
   ? [serviceUnavailable, pageNotFound, serviceProblem]
   : []
