+++ src/schemas/business-details/business-name-change.js
@@ -1,7 +1,5 @@
 import Joi from 'joi'
-import {
-  BUSINESS_NAME_MAX
-} from '../../constants/validation-fields.js'
+import { BUSINESS_NAME_MAX } from '../../constants/validation-fields.js'
 
 export const businessNameSchema = Joi.object({
   businessName: Joi.string()
