+++ src/utils/caching/cache-engine.js
@@ -2,10 +2,10 @@ import { Engine as CatboxRedis } from '@hapi/catbox-redis'
 import { Engine as CatboxMemory } from '@hapi/catbox-memory'
 
 import { buildRedisClient } from './redis-client.js'
-import { config } from '../../config/config.js'
+import { config } from '../../config/index.js'
 import { createLogger } from '../logger.js'
 
-export function getCacheEngine (engine) {
+export const getCacheEngine = (engine) => {
   const logger = createLogger()
 
   if (engine === 'redis') {
@@ -14,7 +14,7 @@ export function getCacheEngine (engine) {
     return new CatboxRedis({ client: redisClient })
   }
 
-  if (config.get('isProduction')) {
+  if (config.get('server.isProduction')) {
     logger.error(
       'Catbox Memory is for local development only, it should not be used in production!'
     )
