+++ src/utils/logger.js
@@ -2,6 +2,6 @@ import { pino } from 'pino'
 
 import { loggerOptions } from '../config/logger-options.js'
 
-export function createLogger () {
+export const createLogger = () => {
   return pino(loggerOptions)
 }
