+++ test/unit/plugins/template-rendering/navigation-items.test.js
@@ -1,7 +1,7 @@
 import { getNavigationItems } from '../../../../src/config/navigation-items.js'
 import { describe, test, expect } from '@jest/globals'
 
-function createMockRequest (testPath) {
+const createMockRequest = (testPath) => {
   return { path: testPath }
 }
